implement builder class and split implementation
